refactor: entity -> domain
