+++ src/main/java/org/myteam/server/news/dto/controller/response/NewsListResponse.java
@@ -1,7 +1,7 @@
 package org.myteam.server.news.dto.controller.response;
 
+import org.myteam.server.global.page.response.PageCustomResponse;
 import org.myteam.server.news.dto.service.response.NewsDto;
-import org.myteam.server.global.page.response.PageCustom;
 
 import lombok.Builder;
 import lombok.Getter;
@@ -11,14 +11,14 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class NewsListResponse {
 
-	private PageCustom<NewsDto> newsList;
+	private PageCustomResponse<NewsDto> newsList;
 
 	@Builder
-	public NewsListResponse(PageCustom<NewsDto> newsList) {
+	public NewsListResponse(PageCustomResponse<NewsDto> newsList) {
 		this.newsList = newsList;
 	}
 
-	public static NewsListResponse of(PageCustom<NewsDto> newsList) {
+	public static NewsListResponse of(PageCustomResponse<NewsDto> newsList) {
 		return NewsListResponse.builder()
 			.newsList(newsList)
 			.build();
