+++ src/main/java/org/myteam/server/news/dto/service/request/NewsServiceRequest.java
@@ -1,6 +1,6 @@
 package org.myteam.server.news.dto.service.request;
 
-import org.myteam.server.news.entity.NewsCategory;
+import org.myteam.server.news.domain.NewsCategory;
 import org.myteam.server.news.repository.OrderType;
 import org.myteam.server.global.page.request.PageInfoServiceRequest;
