+++ src/main/java/org/myteam/server/news/repository/NewsCountRepository.java
@@ -1,6 +1,6 @@
 package org.myteam.server.news.repository;
 
-import org.myteam.server.news.entity.NewsCount;
+import org.myteam.server.news.domain.NewsCount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NewsCountRepository extends JpaRepository<NewsCount, Long> {
