+++ src/main/java/org/myteam/server/news/repository/NewsRepository.java
@@ -1,6 +1,6 @@
 package org.myteam.server.news.repository;
 
-import org.myteam.server.news.entity.News;
+import org.myteam.server.news.domain.News;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface NewsRepository extends JpaRepository<News, Long> {
