+++ src/test/java/org/myteam/server/news/service/NewsReadServiceTest.java
@@ -10,15 +10,15 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.myteam.server.IntegrationTestSupport;
 import org.myteam.server.news.dto.controller.response.NewsListResponse;
-import org.myteam.server.news.entity.News;
-import org.myteam.server.news.entity.NewsCategory;
-import org.myteam.server.news.entity.NewsCount;
+import org.myteam.server.news.domain.News;
+import org.myteam.server.news.domain.NewsCategory;
+import org.myteam.server.news.domain.NewsCount;
 import org.myteam.server.news.repository.NewsCountRepository;
 import org.myteam.server.news.repository.NewsRepository;
 import org.myteam.server.news.repository.OrderType;
 import org.myteam.server.news.dto.service.response.NewsDto;
 import org.myteam.server.news.dto.service.request.NewsServiceRequest;
-import org.myteam.server.global.page.response.PageableCustom;
+import org.myteam.server.global.page.response.PageableCustomResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 
 class NewsReadServiceTest extends IntegrationTestSupport {
@@ -54,7 +54,7 @@ class NewsReadServiceTest extends IntegrationTestSupport {
 		NewsListResponse newsListResponse = newsReadService.findAll(newsServiceRequest);
 
 		List<NewsDto> newsList = newsListResponse.getNewsList().getContent();
-		PageableCustom pageInfo = newsListResponse.getNewsList().getPageInfo();
+		PageableCustomResponse pageInfo = newsListResponse.getNewsList().getPageInfo();
 
 		assertThat(pageInfo)
 			.extracting("currentPage", "totalPage", "totalElement")
@@ -99,7 +99,7 @@ class NewsReadServiceTest extends IntegrationTestSupport {
 		NewsListResponse newsListResponse = newsReadService.findAll(newsServiceRequest);
 
 		List<NewsDto> newsList = newsListResponse.getNewsList().getContent();
-		PageableCustom pageInfo = newsListResponse.getNewsList().getPageInfo();
+		PageableCustomResponse pageInfo = newsListResponse.getNewsList().getPageInfo();
 
 		assertThat(pageInfo)
 			.extracting("currentPage", "totalPage", "totalElement")
@@ -149,7 +149,7 @@ class NewsReadServiceTest extends IntegrationTestSupport {
 		NewsListResponse newsListResponse = newsReadService.findAll(newsServiceRequest);
 
 		List<NewsDto> newsList = newsListResponse.getNewsList().getContent();
-		PageableCustom pageInfo = newsListResponse.getNewsList().getPageInfo();
+		PageableCustomResponse pageInfo = newsListResponse.getNewsList().getPageInfo();
 
 		assertThat(pageInfo)
 			.extracting("currentPage", "totalPage", "totalElement")
@@ -191,7 +191,7 @@ class NewsReadServiceTest extends IntegrationTestSupport {
 		NewsListResponse newsListResponse = newsReadService.findAll(newsServiceRequest);
 
 		List<NewsDto> newsList = newsListResponse.getNewsList().getContent();
-		PageableCustom pageInfo = newsListResponse.getNewsList().getPageInfo();
+		PageableCustomResponse pageInfo = newsListResponse.getNewsList().getPageInfo();
 
 		assertThat(pageInfo)
 			.extracting("currentPage", "totalPage", "totalElement")
